Added protetection in the method call handlers for situations where the controllers have been terminated. This can occure if you start/stop the player/recorder. On start the channel can issues events before the controllers are created.
Implemented a number of explicit exceptions for common cases to make it easy to catch common error conditions (is already playing, is already stopeed).

Moved the isPlaying check for startPlayer and similar as ti was starting the player
adding callbacsk and then throwing an exception which left the callbacks in an inconsistent state.