Update simple_recorder.dart
I used your package and it was great experience. I referred to this example, and I found it had a redundancy and its function do nothing when pressing the buttons. I'm a newbie so I might be wrong. But I think your package is really great, and I just want to contribute to this project if me finding is correct. Thank you guys for sharing this wonderful work.