Improved widget UI
RecorderPlaybackController and the associated UI widgets have been updated with the capability to disable the recorder UI during playback. Testing in the demo app has gone well, but there is an odd behavior where it fails a null-check if you leave and return to the widget demo screen. Investigating root cause.