Disabled player when recording new clip
Made it so if you start recording a new file when an old file has previously been recorded, it will disable the player, thus heading off a bunch of bad behaviors. Ultimately it would be nice to splice into a stream instead of having to re-record, or have the option to add an alert that you are about to erase the existing recording. Also the trash can doesn't work. But this should now be functional and way less buggy.